refactor: tighten decoupling per SOLID (no behaviour change)
Self-audit follow-up — elevate the rearchitecture to the codebase's elegance bar
without reintroducing DI ceremony:

- Ecosystem layer (SRP + remove speculative code): the reflection mechanics that
  talk to the external Euterpe mod move into EuterpeBridgeClient, a named
  anti-corruption layer (type resolution / method caching / out-param marshalling
  / unit conversion all confined there). ChartSource becomes a near-stateless
  pure router (Euterpe -> bridge, else native store).
- Dropped dead code: ChartSource.GetEcosystem, the ChartEcosystem enum, and the
  CustomAlbums detection had NO consumer (CAM correctly uses the native fallback,
  so detecting it changed nothing) — speculative API removed. A strategy pattern
  was deliberately NOT introduced: with Euterpe the only special ecosystem, it
  would be a one-implementation abstraction.
- BattleUi SRP: the self-contained score-zoom state machine moves to
  ScoreZoomFollower; BattleUi keeps overlay construction + the skill-indicator
  offset (left in place because it shares offset state with the upper-left text
  binding — extracting it would create awkward accessor coupling).

Deliberately kept: process-global infrastructure (Log, Config, ...) as ambient
statics — Harmony patches are static entry points and Il2Cpp code can't be unit
tested, so DI's main benefits don't apply; injecting them everywhere is the
over-engineering this rework removed.
